Move AnalyticsConfiguration to analytics module
COSDK-500
