+++ core/src/main/java/com/adyen/checkout/core/CheckoutConfiguration.kt
@@ -13,6 +13,7 @@ import android.os.Parcel
 import android.os.Parcelable
 import android.os.Parcelable.CONTENTS_FILE_DESCRIPTOR
 import androidx.annotation.RestrictTo
+import com.adyen.checkout.core.analytics.AnalyticsConfiguration
 import com.adyen.checkout.core.data.model.Amount
 import com.adyen.checkout.core.internal.Configuration
 import com.adyen.checkout.core.internal.util.CheckoutConfigurationMarker
