+++ core/src/main/java/com/adyen/checkout/core/internal/Configuration.kt
@@ -10,7 +10,7 @@ package com.adyen.checkout.core.internal
 
 import android.os.Parcelable
 import androidx.annotation.RestrictTo
-import com.adyen.checkout.core.AnalyticsConfiguration
+import com.adyen.checkout.core.analytics.AnalyticsConfiguration
 import com.adyen.checkout.core.Environment
 import com.adyen.checkout.core.data.model.Amount
 import java.util.Locale
