+++ core/src/main/java/com/adyen/checkout/core/mbway/internal/ui/MBWayConfiguration.kt
@@ -9,7 +9,7 @@
 package com.adyen.checkout.core.mbway.internal.ui
 
 import android.content.Context
-import com.adyen.checkout.core.AnalyticsConfiguration
+import com.adyen.checkout.core.analytics.AnalyticsConfiguration
 import com.adyen.checkout.core.CheckoutConfiguration
 import com.adyen.checkout.core.Environment
 import com.adyen.checkout.core.data.model.Amount
